Added display for details, and edited the loop to simulate a clock cycle
Clock cycle is simulated such that in every cycle we have a random number from 0 to n-1 generated where n is the # of cores. So, each peice of data goes into a random core and is processed at each simulated cycle. 0 to n-1 because the indeces of the array holding the core objects is in that format.